refactor: add HTTPProviderResource base
